修改了部门号的数据类型为Integer

+++ src/main/java/com/tedu/sois/common/config/StatusCodeConfig.java
@@ -22,7 +22,7 @@ public interface StatusCodeConfig {
     String REGEX_EX = "[`~!@#$%^&*()+=|{}':;',\\[\\].<>/?~！@#￥%……&*（）——+|{}【】‘；：”“’。，、？]";
 
     /** 学生部门号*/
-    Long STU_DEPT_CODE = 13L;
+    int STU_DEPT_CODE = 13;
